Refactor worker into modular handlers
