+++ src/utils/random.ts
@@ -1,0 +1,8 @@
+export function urlSafeRandom(len = 12): string {
+  const bytes = new Uint8Array(len);
+  crypto.getRandomValues(bytes);
+  const alph = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
+  let out = '';
+  for (let i = 0; i < len; i++) out += alph[bytes[i] % alph.length];
+  return out;
+}
